+++ src/ninja/mbedded/ninjaterm/util/javafx/comDataPaneWeb/ComDataPaneWeb.java
@@ -90,7 +90,7 @@ public class ComDataPaneWeb extends StackPane {
 
     private SimpleObjectProperty<ScrollState> scrollState = new SimpleObjectProperty<>(ScrollState.FIXED_TO_BOTTOM);
 
-    private int currNumChars = 0;
+    public SimpleIntegerProperty currNumChars = new SimpleIntegerProperty(0);
 
     private WebEngine webEngine;
 
@@ -166,9 +166,10 @@ public class ComDataPaneWeb extends StackPane {
 
         bufferSize = new SimpleIntegerProperty(DEFAULT_BUFFER_SIZE);
         bufferSize.addListener((observable, oldValue, newValue) -> {
+            // If the buffer size is changed, we may need to trim the data
+            // to fit the new size (if smaller)
             trimIfRequired();
         });
-
 
         //==============================================//
         //================= SCROLL SETUP ===============//
@@ -285,7 +286,7 @@ public class ComDataPaneWeb extends StackPane {
         appendHtml(html);
 
         // Update the number of chars added with what was added to the last existing text node
-        currNumChars += textToAppend.toString().length();
+        currNumChars.set(currNumChars.get() + textToAppend.toString().length());
 
 
         // Create new text nodes and copy all text
@@ -333,7 +334,7 @@ public class ComDataPaneWeb extends StackPane {
             appendHtml(html);
 
             // Update the num. chars added with all the text added to this new Text node
-            currNumChars += textToAppend.length();
+            currNumChars.set(currNumChars.get() + textToAppend.toString().length());
         }
 
         if (streamedData.getColorToBeInsertedOnNextChar() != null) {
@@ -422,7 +423,7 @@ public class ComDataPaneWeb extends StackPane {
         currScrollPos = 0;
 
         // Reset character count
-        currNumChars = 0;
+        currNumChars.set(0);
 
 
     }
@@ -434,17 +435,17 @@ public class ComDataPaneWeb extends StackPane {
 
         logger.debug("trimIfRequired() called.");
 
-        if (currNumChars >= bufferSize.get()) {
+        if (currNumChars.get() >= bufferSize.get()) {
 
-            int numCharsToRemove = currNumChars - bufferSize.get();
+            int numCharsToRemove = currNumChars.get() - bufferSize.get();
             logger.debug("Need to trimIfRequired display text. currNumChars = " + currNumChars + ", numCharsToRemove = " + numCharsToRemove);
 
             webEngine.executeScript("trim(" + numCharsToRemove + ")");
 
             // Update the character count
-            currNumChars = currNumChars - numCharsToRemove;
+            currNumChars.set(currNumChars.get() - numCharsToRemove);
 
-            logger.debug("currNumChars = " + currNumChars);
+            logger.debug("currNumChars.get() = " + currNumChars.get());
         }
     }